requirements.txt
pytest=6.2.4
requests=2.26.0
vcrpy=4.1.1